% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_05_user_roles.R
\name{get_user_roles}
\alias{get_user_roles}
\title{noClocksAuthR API - Get User Role(s)}
\usage{
get_user_roles(user_uid = NULL, role_uid = NULL, api_key = get_api_key())
}
\arguments{
\item{user_uid}{an optional user uid.}

\item{role_uid}{an optional role uid.}

\item{api_key}{your noClocksAuthR API key.  Set your noClocksAuthR api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{noclocksauthr_api_res}.  The "content" of the object is a
tibble of users(s) with the following columns:
\itemize{
\item role_uid
\item role_name,
\item user_uid,
\item user_name,
\item created_at
}
}
\description{
noClocksAuthR API - Get User Role(s)
}
\seealso{
\code{\link[=add_user_role]{add_user_role()}} \code{\link[=delete_user_role]{delete_user_role()}}
}
