% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_00.R
\name{noclocksauthr_api_res}
\alias{noclocksauthr_api_res}
\title{Send GET Request to the noClocksAuthR API}
\usage{
noclocksauthr_api_res(resp)
}
\arguments{
\item{resp}{a noClocksAuthR API response}
}
\value{
an S3 object of class "noclocksauthr_api_res".
}
\description{
Send GET Request to the noClocksAuthR API
}
