% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_00.R
\name{print.noclocksauthr_api_res}
\alias{print.noclocksauthr_api_res}
\title{print noclocksauthr_api_res}
\usage{
\method{print}{noclocksauthr_api_res}(x, ...)
}
\arguments{
\item{x}{an S3 object of class \code{noclocksauthr_api_res}.}

\item{...}{additional arguments.}
}
\value{
\code{invisible(NULL)}
}
\description{
Generic print function for \code{noclocksauthr_api_res} S3 class.
}
