% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_00.R
\name{set_api_key}
\alias{set_api_key}
\alias{get_api_key}
\title{set noClocksAuthR API key}
\usage{
set_api_key(api_key)

get_api_key()
}
\arguments{
\item{api_key}{the noClocksAuthR API key}
}
\value{
a list of the newly set \code{noClocksAuthR} R options
}
\description{
The API key can be set as an Environment Variable via
\code{Sys.getenv("NOCLOCKSAUTHR_API_KEY")}.
}
\examples{

set_api_key(api_key = "<my noClocksAuthR API key>")


}
