% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_02_users.R
\name{update_user}
\alias{update_user}
\title{noClocksAuthR API - Update a user}
\usage{
update_user(user_uid, user_data, api_key = get_api_key())
}
\arguments{
\item{user_uid}{the uid of the user to be updated.}

\item{user_data}{list of data to update.}

\item{api_key}{your noClocksAuthR API key.  Set your noClocksAuthR api key using \code{\link{set_api_key}()}
so that you do not need to supply this argument with each function call.}
}
\value{
an object of class \code{noclocksauthr_api_res}.  When successful, the \code{content} of the
\code{noclocksauthr_api_res} is \code{list(message = "success")}.  In the case of an error, the
content is \code{list(error = "<error message>")}.
}
\description{
noClocksAuthR API - Update a user
}
\seealso{
\code{\link[=get_users]{get_users()}} \code{\link[=add_user]{add_user()}} \code{\link[=delete_user]{delete_user()}}
}
